# prism/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API gateway with service routing, audit logging, API key management, and service discovery.",
        "title": "Prism API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/audit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated, filtered list of audit logs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "List audit logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by tenant ID",
                        "name": "tenant_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by event type",
                        "name": "event_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by resource type",
                        "name": "resource_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by resource ID",
                        "name": "resource_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuditLogListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/audit/aggregate/{groupBy}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Group and count audit logs by a specified field",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Aggregate audit logs",
                "parameters": [
                    {
                        "enum": [
                            "event_type",
                            "actor_type",
                            "tenant"
                        ],
                        "type": "string",
                        "description": "Group by field",
                        "name": "groupBy",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuditLogAggregateResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/audit/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Export audit logs as CSV file",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Export audit logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by tenant ID",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/audit/{auditID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single audit log entry",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Get audit log by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit log ID",
                        "name": "auditID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuditLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid JSON request"
                },
                "slug": {
                    "type": "string",
                    "example": "invalid-json"
                }
            }
        },
...
        "AuditLog": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "description": "ActorID is the unique identifier of the actor who performed the action such as a user or api key (usr_1234567890, apk_1234567890)",
                    "type": "string"
                },
                "actor_type": {
                    "description": "ActorType identifies the type of entity that performed the action (e.g., \"user\", \"system\", \"service\")",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt is the timestamp when the audit log entry was created",
                    "type": "string"
                },
                "event_data": {
                    "description": "EventData stores additional contextual information about the event in a flexible format"
                },
                "event_type": {
                    "description": "EventType categorizes the type of action performed (e.g., \"login\", \"create\", \"update\", \"delete\")",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the unique identifier for the audit log entry with the \"aud\" prefix"
                },
                "ip_address": {
                    "description": "IPAddress stores the IP address from which the action was performed",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "resource_id": {
                    "description": "ResourceID is the unique identifier of the resource that was affected with the \"res\" prefix"
                },
                "resource_type": {
                    "description": "ResourceType identifies the type of resource that was affected (e.g., \"user\", \"apikey\", \"service\")",
                    "type": "string"
                },
                "tenant_id": {
                    "description": "TenantID is the unique identifier of the tenant where the action occurred with the \"tnt\" prefix"
                },
                "user_agent": {
                    "description": "UserAgent contains the user agent string from the client that performed the action",
                    "type": "string"
                }
            }
        },
        "AuditLogAggregate": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "group_key": {
                    "type": "string"
                }
            }
        },
        "AuditLogAggregateResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AuditLogAggregate"
                    }
                },
                "end_date": {
                    "type": "string"
                },
                "group_by": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "AuditLogListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AuditLog"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/PaginationMetadata"
                }
            }
        },
...
        "PaginationMetadata": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}